requirements.txt
streamlit==1.56.0
PyYAML==6.0.2
pandas==2.3.3
requests==2.33.1
beautifulsoup4==4.14.3
playwright==1.58.0
playwright-stealth==2.0.3
pymysql==1.1.2
openpyxl==3.1.5
xlrd==2.0.2
